+++ stdlib/LinearAlgebra/test/svd.jl
@@ -82,7 +82,7 @@ a2img  = randn(n,n)/2
             β = svdfact(α)
             @test β.S == [abs(α)]
             @test svdvals(α) == abs(α)
-            u,v,q,d1,d2,r0 = svd(a,a_svd)
+            u,v,q,d1,d2,r0 = svdfact(a,a_svd)
             @test u ≈ gsvd.U
             @test v ≈ gsvd.V
             @test d1 ≈ gsvd.D1
@@ -103,10 +103,8 @@ a2img  = randn(n,n)/2
             x, y = randn(eltya, 2)
             @test svdfact(x)    == svdfact(fill(x, 1, 1))
             @test svdvals(x)    == first(svdvals(fill(x, 1, 1)))
-            @test svd(x)        == first.(svd(fill(x, 1, 1)))
             @test svdfact(x, y) == svdfact(fill(x, 1, 1), fill(y, 1, 1))
             @test svdvals(x, y) ≈  first(svdvals(fill(x, 1, 1), fill(y, 1, 1)))
-            @test svd(x, y)     == first.(svd(fill(x, 1, 1), fill(y, 1, 1)))
         end
     end
     if eltya != Int